moved ProblemManager's threading logic to a new place : ThreadedRangeProcessor
